rate変数の型をdoubleからlongに変更中 ビット演算をしやすい

+++ src/qow/framework/system/FrameLoop.java
@@ -19,7 +19,7 @@ public class FrameLoop extends Loop {
      *
      * @param rate レートの初期値
      */
-    protected FrameLoop(double rate) {
+    protected FrameLoop(long rate) {
         this();
         setRate(rate);
     }
